+++ src/ya_tagscript/blocks/strings/__init__.py
@@ -1,5 +1,6 @@
 from .case_block import CaseBlock
 from .join_block import JoinBlock
+from .python_block import PythonBlock
 from .replace_block import ReplaceBlock
 from .substring_block import SubstringBlock
 from .urldecode_block import URLDecodeBlock
@@ -8,6 +9,7 @@ from .urlencode_block import URLEncodeBlock
 __all__ = (
     "CaseBlock",
     "JoinBlock",
+    "PythonBlock",
     "ReplaceBlock",
     "SubstringBlock",
     "URLDecodeBlock",